removed job package from upload location when the job is killed
